srcs/requirements/requirements.txt
asgiref==3.8.1
Cython==3.0.11
Django==4.1.13
django-cors-headers==4.4.0
djongo==1.3.6
dnspython==2.6.1
pillow==10.4.0
pymongo==3.12.3
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
sqlparse==0.2.4
typing_extensions==4.12.2
